+++ src/main/java/com/teamtea/eclipticseasons/data/loot/EclipticSeasonsLootTableProvider.java
@@ -1,13 +1,11 @@
 package com.teamtea.eclipticseasons.data.loot;
 
-import net.minecraft.core.HolderLookup;
 import net.minecraft.data.PackOutput;
 import net.minecraft.data.loot.LootTableProvider;
 import net.minecraft.world.level.storage.loot.parameters.LootContextParamSets;
 
 import java.util.List;
 import java.util.Set;
-import java.util.concurrent.CompletableFuture;
 
 
 public class EclipticSeasonsLootTableProvider extends LootTableProvider {