+++ src/main/java/com/teamtea/eclipticseasons/data/tag/ESItemTagProvider.java
@@ -4,13 +4,16 @@ package com.teamtea.eclipticseasons.data.tag;
 import com.teamtea.eclipticseasons.api.EclipticSeasonsApi;
 import com.teamtea.eclipticseasons.api.constant.crop.CropHumidityType;
 import com.teamtea.eclipticseasons.api.constant.crop.CropSeasonType;
+import com.teamtea.eclipticseasons.api.constant.tag.ESItemTags;
 import com.teamtea.eclipticseasons.api.constant.tag.EclipticBlockTags;
 import net.minecraft.core.HolderLookup;
 import net.minecraft.data.PackOutput;
 import net.minecraft.data.tags.ItemTagsProvider;
 import net.minecraft.resources.ResourceLocation;
+import net.minecraft.tags.ItemTags;
 import net.minecraft.world.item.Items;
 import net.minecraft.world.level.block.Block;
+import net.minecraftforge.common.Tags;
 
 import java.util.concurrent.CompletableFuture;
 
@@ -24,7 +27,7 @@ public final class ESItemTagProvider extends ItemTagsProvider {
 
     @Override
     public String getName() {
-        return EclipticSeasonsApi.MODID+ " Item Tags";
+        return EclipticSeasonsApi.MODID + " Item Tags";
     }
 
     @Override
@@ -42,10 +45,10 @@ public final class ESItemTagProvider extends ItemTagsProvider {
         tag(CropHumidityType.DRY_AVERAGE.getTag()).add(Items.CACTUS);
         tag(CropHumidityType.DRY_MOIST.getTag()).add(Items.SWEET_BERRIES);
         tag(CropHumidityType.DRY_HUMID.getTag()).add(Items.MELON_SEEDS);
-        tag(CropHumidityType.AVERAGE_HUMID.getTag()).add(Items.GLOW_BERRIES,Items.SUGAR_CANE);
+        tag(CropHumidityType.AVERAGE_HUMID.getTag()).add(Items.GLOW_BERRIES, Items.SUGAR_CANE);
         tag(CropHumidityType.AVERAGE_MOIST.getTag()).add(Items.WHEAT_SEEDS, Items.CARROT, Items.BEETROOT_SEEDS, Items.POTATO, Items.PUMPKIN_SEEDS);
         tag(CropHumidityType.AVERAGE_MOIST.getTag()).add(Items.COCOA_BEANS, Items.KELP, Items.TORCHFLOWER_SEEDS);
-        tag(CropHumidityType.MOIST_HUMID.getTag()).add(Items.BAMBOO).add(Items.BROWN_MUSHROOM,Items.RED_MUSHROOM);
+        tag(CropHumidityType.MOIST_HUMID.getTag()).add(Items.BAMBOO).add(Items.BROWN_MUSHROOM, Items.RED_MUSHROOM);
 
         // others
         tag(CropHumidityType.AVERAGE_MOIST.getTag()).addOptional(fd_rl("tomato_seeds")).addOptional(fd_rl("cabbage_seeds")).addOptional(fd_rl("onion"));
@@ -58,6 +61,9 @@ public final class ESItemTagProvider extends ItemTagsProvider {
             tag(cropHumidityType.getTag());
         }
 
+
+        tag(ESItemTags.COOLING_ITEMS).add(Items.SNOWBALL, Items.SNOW_BLOCK, Items.ICE, Items.BLUE_ICE, Items.PACKED_ICE);
+        tag(ESItemTags.HEAT_PROTECTIVE_HELMETS);
     }